findAll & save fonctionnels

+++ pizzeria-app/pizzeria-dao/src/test/java/fr/pizzeria/model/PizzaDaoImplTest.java
@@ -1,15 +1,17 @@
 package fr.pizzeria.model;
 
-import org.junit.*;
-
-import fr.pizzeria.dao.PizzaDaoImpl;
-import fr.pizzeria.model.CategoriePizza;
-import fr.pizzeria.model.Pizza;
-
-import static org.hamcrest.CoreMatchers.*;
-import static org.junit.Assert.*;
+import static org.hamcrest.CoreMatchers.anyOf;
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 
 import java.util.List;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import fr.pizzeria.dao.PizzaDaoImpl;
 
 public class PizzaDaoImplTest {
 	PizzaDaoImpl dao;
@@ -27,9 +29,9 @@ public class PizzaDaoImplTest {
 
 	@Test
 	public void testEquals() {
-		Pizza p1 = new Pizza("GRE", "gredue", 0, CategoriePizza.VIANDE);
+		Pizza p1 = new Pizza("GRE", "gredue", (double) 0, CategoriePizza.VIANDE);
 		Pizza p2 = new Pizza("GRE", "grebel", 4.5, CategoriePizza.VEGETARIEN);
-		Pizza p3 = new Pizza("GAB", "gabrie", 0, CategoriePizza.VIANDE);
+		Pizza p3 = new Pizza("GAB", "gabrie", (double) 0, CategoriePizza.VIANDE);
 		assertTrue(p1.equals(p1));
 		assertTrue(p1.equals(p2));
 		assertFalse(p1.equals(p3));
